Added connection object (a sample connection table is added to the SVN here) Successfully matches against another connection table, and returns whether the other table is a subset of the main one.
We will use this to check that submitted h5 files have a connection table that is valid within the context of the current experiment connection table. Ones that don't pass the test will be rejected.

This protects somewhat against people submitting jobs to the wrong machine, or trying to submit jobs to a machine that isn't configured for their experiment.