Appending data at end of acquisition is now very fast!
Writing to h5file is the slowest part.

Can handle 32 channels @ 20kHz for 120s.
Can't handle the same for 180s. I catch an exception, but there is no error message. Might be to do with the size of the numpy array? May have to split, and write consecutive numpy arrays to the h5 file