You can now pass pickleable or unpickleable objects to the worker class at its initialisation, by passing a dictionary workerargs, for eg workerargs={'x':Queue(),'y':'foobar'} to the Tab class initialisation.
The worker object will then have x and y added as instance variables self.x and self.y.

This was motivated by the requirement to pass an extra Queue object to the worker process of the NI card.

Also if you do try to pass unpickleable objects to the worker later, you used to get a hard crash. Now the worker simply stops responding, and a reset works fine. I don't seem to be able to catch the exception -- it doesn't appear to happen in either the parent or child exceptions! So I'll make this better at some point by checking first that all objects are pickleable before passing them.