Added some more robustness to the programming of hardware. If instruction sets are missing, it shouldn't crash now.
We also now only call transition_to_buffered on devices which are in the submitted h5 file. This should also help cut down on device crashes!

If we do crash a tab, or a tab is already in an error state, the transition_to_buffered will be aborted, the h5 placed back at the head of the queue, and the queue paused until the user intervenes.