Resolves #476 refs #345
The NovatechDDS9m now compares front panel values to device values and gives the user the opportunity to choose on a per channel basis which values they want (either front panel or device).

This also makes it easy to restore the Novatech channel lock state, units, step size as I don't have to write a special case to do it!

This feature also makes it easy to restore values to your novatech if it loses power!

The front panel is greyed out while waiting for the user to decide which values to accept.
Running a buffered experiment while the front panel is greyed out is fine, it just ignores both the remote and front panel values as they'll be overwritten at the end of the run anyway!

The pretend "gate" has also been removed as we should have most channels under digital gate control anyway!