In output_classes.py Made function names consistent across AO/DO objects Added functions to the DDS object to get/set the value. Values are dictionaries containing channel:value pairs
In tab_base_classes.py
Cleaned up the code a bit to remove lots of redundant GTK stuff
Fixed error reporting (made it a lot cleaner, and now it is formatted and coloured correctly)
Adjusted the code to reflect the changes to the UI file

In tab_frame.ui
Made the title, restart and refresh smart programming widgets part of every tab
Will make it possible to hide the smart programming if the device doesn't use it (maybe disable and force check the box for consistency?)

In device_base_class.py (ALL UNTESTED)
Wrote a class to try and abstract away a lot of the boring stuff we do with the UI in every device GUI class.
Widget generation is more automatic, and it should even be able to place widgets for you.
Channels and their capabilities are defined in a dictionary (kind of like a definition through meta data)
Added standard functions for transitioning between modes, etc.
I expect most devices to subclass this, and at a minimum, overwrite the initialise_GUI() method, which should call some of the widget generation classes previously described. A worker class can then be made to match the standard worker process calls I've used here (like transition_to_manual, etc)