renamed state to mode in some instances to make the distinction between states and modes clearer!
Added a delete_stale_states parameter to @define_state which will pull the most recent state out of the queue of adjacent identical statefunctions.

Eg, if the statefunctions in the queue are: [a,a,a,b,a,a,c,a,a,a,a]
and a is defined wth delete_stale_states=True,
then we will execute the 3rd a, then b, then skip the next a, execute the a before the c, execute c, and then execute the last a.
This is designed so that multiple calls to program_device only runs the most recent one, which is especially important on slow devices like the RfBlaster which can take ~5s to program in manual mode. This replaces the hack previously used in the RfBlaster which ignored calls to program_device if we already had 2 in the queue (the number in the queue was kept track of manually within the tab)