1.1.0-dev: Fixed race condition in accessing self.last_requested_state
This probably didn't affect much, but had at least a small probability of raising an exception. The exception occurred when Python switched threads half-way through evaulating "if self.last_requested_state is not None and allowed_states&self.last_requested_state:" which raised an exception if self.last_requested_state was None and "&"'d with an int.