1.1.0-dev: Renamed icon to lower case.
I did this earlier (just by renaming the file) but mercurial didn't pick
up on it because windows filenames are not case sensitive.

However windows *is* case sensitive enough that it couldn't find the icon file
later when I set it on a shortcut with the new name. Silly....

Also removed a pointless printline