callback_priorities: Added priorities for plugin callbacks. Plugin callbacks may now be decorated with the decorator `blacs.plugins.callback(priority=<number>)`, and those with the lowest number for their priority will be called first. The defauly priority is 10, which will be used if no priority is specified or if a callback does not use this decorator. Because it is optional, this change is backward compatible with existing callbacks, which will work unmodified. I have not added the decorator to any of the callbacks of the plugins shipped with BLACS.
This functionality was implemented by defining a Callback class in blacs.plugins
that has an attribute for the priority, but is also a callable, and so can be
just called as normal if the priority is not relevant.

Moved `get_callbacks()` from `blacs.experiment_queue` into `blacs.plugins`, and
modified it to return callbacks in order of priority. Removed the `inmain`
decorator from it since it should not normally need to touch the GUI - if a
plugin needs to touch the GUI to get callbacks for whatever reason, it can call
`inmain()` itself.

Replaced the remaining code in the
queue manager that manually looked through plugins for callbacks with
a call to get_callbacks.

This is so that I can make a plugin with a callback for 'shot_complete', which
delays the start of the next shot. This callback should run last if there are
other callbacks, necessitating priorities of some sort.