Move config path to same location as other apps
And have it named BLACS.h5 instead of containing the hostname.

This ensures we don't use variables no longer defined in
`labscript_utils.labconfig` as of labscript-suite/labscript-utils#45,
whilst also conforming to a more standard layout of where the config
file is stored.

A nice side effect of storing the BLACS config in the app_saved_configs
dir is that if the user changes the experiment name in their labconfig
file, they can switch between BLACS configs from different experiments
without having to recompile the connection table, change globals files
in BLACS, etc. So different experiments on the same machine can have a
more self-contained config.

I don't see much upside to keeping the hostname in the config filename,
so now it's just BLACS.h5. With labconfig there is at least the idea
that you might version control the labconfig directory to share changes
between computers, but that is not as useful for BLACS' config.