Retry attempts on communication between parent and child
Added in try blocks around get()/put() commands between worker and its parent tab. A timeout value of 30 seconds is added and a TimeoutError exception is captured to retry the message.

Reason: When using secure communication zmq messages get lost/dropped/missed between the parent tab and its worker seemingly at random. The most often scenario of this occurring is when sending a new job to a worker. The parent will issue a job to the worker and proceed to wait for an acknowledgement. The worker never receives this message and continues to wait for a job from the parent. This causes the blacs tab to enter an infinite wait state, and the tab will error out with a 'Device has not responded for xx:xx time'. The tab will then sit in this error state counting up for hours, days, or weeks until the tab is restarted. Bringing any experiment to a halt. This seems to occur infrequently, but devices which send requests repeatedly in a short amount of time tend to encounter this error more often (such as the Pulseblaster with its 'check_status' job.)

This error is not computer specific, and seems to plague all computers on the network, even those simply communicating over their localhost loopback interface. Adding in these retry blocks seem to get around this issue if turning off secure communication is not an option. So far only one retry is needed to restore communication between the tab and its worker.